refactor: custom paths for signup db functions
