+++ package.json
@@ -17,7 +17,6 @@
 		"pg": "^8.11.3",
 		"prisma": "^5.0.0",
 		"string-similarity": "^4.0.4",
-		"tsconfig-paths": "^4.2.0",
 		"zod": "^3.22.1"
 	},
 	"devDependencies": {
@@ -29,7 +28,8 @@
 		"@typescript-eslint/parser": "^6.4.1",
 		"drizzle-kit": "^0.19.13",
 		"eslint": "^8.48.0",
-		"ts-node": "^10.9.1"
+		"ts-node": "^10.9.1",
+		"tsconfig-paths": "^4.2.0"
 	},
 	"nodemonConfig": {
 		"ignore": [
