+++ src/clock/turbos.ts
@@ -1,8 +1,9 @@
 import { ChannelType, type Client, type Guild } from 'discord.js';
 import { sendErrorLog } from '../structures/logs';
 import { client, prisma } from '..';
-import { type FullSignup, createAutomatedGame, getOrCreateUser, getSignup } from '../util/database';
+import { createAutomatedGame, getOrCreateUser } from '../util/database';
 import { formatSignupEmbed } from '../util/embeds';
+import { getSignup, type FullSignup } from '@models/signups';
 
 export const turboExpirySeconds = 60 * 60 * 3;
