+++ src/interactions/commands/playerchats.ts
@@ -1,6 +1,6 @@
 import { ChannelType, type OverwriteResolvable, SlashCommandBuilder, type User } from 'discord.js';
 import { ServerType, newSlashCommand } from '../../structures/BotClient';
-import { getSignup } from '../../util/database';
+import { getSignup } from '@models/signups';
 
 const data = new SlashCommandBuilder().setName('playerchats').setDescription('Create a category as associated channels for a signup');
 data.addStringOption((x) => x.setName('name').setDescription('The name of the game').setRequired(true));
