+++ src/structures/turbos/turboSignups.ts
@@ -1,7 +1,8 @@
 import { type Guild, ChannelType, Colors } from 'discord.js';
 import { client, prisma } from '../..';
-import { type FullSignup, createAutomatedGame } from '../../util/database';
 import { formatSignupEmbed } from '../../util/embeds';
+import { type FullSignup } from '@models/signups';
+import { createAutomatedGame } from '@utils/database';
 
 async function getMessage(guild: Guild, channelId: string, messageId: string) {
 	try {
